requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
cryptography
jinja2
python-multipart
python-dotenv